refactor: move S3StagingConfiguration back to redshift-v2 and rollback CDK changes
