+++ airbyte-integrations/connectors/destination-redshift-v2/gradle.properties
@@ -1,3 +1,3 @@
-cdkVersion=local
+cdkVersion=1.0.7
 testExecutionConcurrency=4
 JunitMethodExecutionTimeout=15m
